Rename dbschemaexport files to dbexportschema for consistency
