+++ src/project/CMakeLists.txt
@@ -30,8 +30,8 @@ qt_add_executable(SQLiteQueryAnalyzer WIN32 MACOSX_BUNDLE
         ../database/dbexport.cpp ../database/dbexport.h
         ../threading/cancellation.cpp ../threading/cancellation.h
         ../threading/mainthread.h
-        ../database/dbschemaexport.cpp
-        ../database/dbschemaexport.h
+        ../database/dbexportschema.cpp
+        ../database/dbexportschema.h
         ../database/dbexportdata.cpp
         ../database/dbexportdata.h
 )
