+++ src/project/SQLiteAnalyzer.pro
@@ -26,7 +26,7 @@ SOURCES += \
     ../settings/recentfiles.cpp \
     ../settings/settings.cpp \
     ../database/dbexport.cpp \
-    ../database/dbschemaexport.cpp \
+    ../database/dbexportschema.cpp \
     ../database/dbexportdata.cpp \
     ../threading/cancellation.cpp
 
@@ -41,7 +41,7 @@ HEADERS += \
     ../settings/recentfiles.h \
     ../settings/settings.h \
     ../database/dbexport.h \
-    ../database/dbschemaexport.h \
+    ../database/dbexportschema.h \
     ../database/dbexportdata.h \
     ../threading/cancellation.h \
     ../threading/mainthread.h
