+++ src/ui/mainwindow.cpp
@@ -3,7 +3,7 @@
 #include "ui_mainwindow.h"
 #include "../settings/settings.h"
 #include "../database/dbexport.h"
-#include "../database/dbschemaexport.h"
+#include "../database/dbexportschema.h"
 #include "../threading/mainthread.h"
 
 #include <QMessageBox>
